refactor: clean up utils
